refactor(session): 提取并迁移会话相关工具函数
- 将会话相关的工具函数从 session/index.ts 中移除
- 在 session/utils.ts 中重新实现并导出这些工具函数
- 调整了导入路径，改用新的工具函数模块
- 修正了多个文件中对 executeValidatedTool 的导入路径
- 统一了 SkillsDropdown 组件中的 DropdownMenu 导入名称
